requirements.txt
python-dotenv
requests
mysql-connector-python==8.0.5